# Main dependencies upgrades with more inlines