--- a/Cargo.toml
+++ b/Cargo.toml
@@ -26,9 +26,9 @@ hashbrown = "0.11.2"
 enum-map = { version = "1.1.1", optional = true }
 
 [dev-dependencies]
-serde_json = "1.0.73"
+serde_json = "1.0.39"
 bencher = "0.1.5"
-proptest = "1.0.0"
+proptest = "0.9.1"
 
 [features]
 dev = []