Out of curiosity @greyblake, I gave it another go and upgraded hashmap and enum-map to the latest releases, and left the dev-dependencies alone.

After sprinkling more `#[inline]`s, the performance is a bit better.

```
running 2 tests
test bench_detect        ... bench:   8,508,715 ns/iter (+/- 571,452)
test bench_detect_script ... bench:     238,753 ns/iter (+/- 17,187)

running 2 tests
test bench_detect        ... bench:   8,378,025 ns/iter (+/- 856,485)
test bench_detect_script ... bench:     237,377 ns/iter (+/- 12,368)

running 2 tests
test bench_detect        ... bench:   8,408,746 ns/iter (+/- 738,716)
test bench_detect_script ... bench:     236,423 ns/iter (+/- 44,616)
```
Creating this PR just in case you want to bump the main dependencies to the latest maintained releases without the huge performance regression.