@greyblake 2022-04-15
## Before

```
test bench_detect        ... bench:   9,065,607 ns/iter (+/- 303,908)
test bench_detect_script ... bench:     113,341 ns/iter (+/- 6,817)

test bench_detect        ... bench:   8,944,207 ns/iter (+/- 315,354)
test bench_detect_script ... bench:     113,169 ns/iter (+/- 5,634)

test bench_detect        ... bench:   9,046,893 ns/iter (+/- 284,733)
test bench_detect_script ... bench:     111,744 ns/iter (+/- 13,799)
```

## After

```
test bench_detect        ... bench:   8,350,379 ns/iter (+/- 213,098)
test bench_detect_script ... bench:     106,073 ns/iter (+/- 22,040)

test bench_detect        ... bench:   8,382,447 ns/iter (+/- 279,420)
test bench_detect_script ... bench:     109,724 ns/iter (+/- 9,798)

test bench_detect        ... bench:   8,382,447 ns/iter (+/- 279,420)
test bench_detect_script ... bench:     109,724 ns/iter (+/- 9,798)
```

What makes ~6.8% performance improvement for `bench_detect`.

@jqnatividad Thanks!